fix(cloud): extract post-debit hint republish to break a module-graph regression
Pass-A review of #17745 found the new `inference-billing-ledger` ->
`inference-billing-fast-path` import edge broke an unrelated existing suite:

  SyntaxError: Export named 'dbRead' not found in module
  packages/cloud/shared/src/db/helpers.ts

`inference-billing-invalidation-logged.test.ts` mocks `../../db/helpers` with
just `{ dbWrite, writeTransaction }`, which was complete for the ledger's old
module graph. Importing the fast path dragged in `api-keys` ->
`db/repositories` -> `dbRead`, so the mock no longer satisfied the graph and the
whole batch aborted. This was a real CI failure on the PR head (`unit-tests`,
batch 6/12), not flake.

Symptom-level fixes rejected: widening the mock leaves a layering violation
where the DB ledger transitively depends on the KV fast path (and `api-keys`)
for one helper. Fixed at the root instead by extracting
`republishOrgBalanceHintAfterDebit` into its own leaf module
`inference-balance-republish.ts`, whose only dependencies are `credits`,
`inference-auth-cache`, and the isolate-local `inference-admission-refusal` —
all of which the ledger already carried. The fast path re-exports the symbol so
existing importers are unchanged.

The invalidation-logged test's induced failure moves from
`invalidateOrgBalanceHint` to `republishOrgBalanceHint`: the same `balance-hint`
cache target reached through the settler's new call. Its assertion (a failing
balance-hint repair still WARNs rather than being swallowed) is unchanged, and
the mocked authoritative read succeeds so the failure under test is
unambiguously the cache write. No check weakened.

Verified: fast-path 42/42 pass; invalidation-logged 1/1 pass;
`packages/cloud/shared` typecheck exits 0; biome clean.
